moved friend_code to a separate module
