+++ animal_crossing.py
@@ -1,5 +1,4 @@
 from datetime import datetime, time, timezone
-import re
 import sqlite3
 import dateutil
 import dateutil.tz
@@ -14,7 +13,6 @@ if config.bot.acnh_db is not None:
 	with db:
 		db.execute('PRAGMA foreign_keys = ON')
 
-friend_code_regex = re.compile(r'SW-\d{4}-\d{4}-\d{4}')
 cached_row_id = 0
 
 def stalk_market(cmd):
@@ -114,80 +112,7 @@ See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for a complete 
 
 	if cur.rowcount:
 		current_time = datetime.now().astimezone(tz)
-		cmd.reply('''Timezone successfully updated.
-Your current time should be %s.
-		''' % (current_time))
+		cmd.reply('Timezone successfully updated. Your current time should be %s.'
+			% (current_time))
 	else:
 		cmd.reply('Timezone could not be updated. Have you registered a friend code?')
-
-def friend_code(cmd):
-	if not cmd.args:
-		_user_list_all(cmd)
-		return
-
-	split = cmd.args.split(' ', 1)
-	subcmd = split[0]
-	if subcmd == 'set':
-		_user_upsert_friend_code(cmd, split)
-	elif subcmd == 'remove':
-		_user_remove(cmd)
-	else:
-		_user_find(cmd, split[0])
-
-def _user_upsert_friend_code(cmd, split):
-	if len(split) != 2:
-		cmd.reply('usage: !fc set friend-code')
-		return
-
-	if friend_code_regex.match(split[1]) is None:
-		cmd.reply('Invalid friend code submitted.')
-		return
-
-	sender = cmd.sender
-	cur = None
-	with db:
-		cur = db.execute('''
-		INSERT INTO user VALUES(?, ?, ?, null)
-		ON CONFLICT(id)
-		DO UPDATE SET code=excluded.code
-		''', (sender['id'], sender['username'], split[1]))
-
-	if cur.rowcount:
-		cmd.reply('Friend code for %s has been set.' % (sender['username']))
-	else:
-		cmd.reply('Could not create user.')
-
-def _user_list_all(cmd):
-	cur = db.execute('SELECT username, code FROM user')
-
-	users = cur.fetchall()
-	if users:
-		reply = []
-		for user in users:
-			reply.append('%s: %s' % (user['username'], user['code']))
-		cmd.reply('\n'.join(reply))
-	else:
-		cmd.reply('There are no friend codes saved.')
-
-def _user_find(cmd, user):
-	cur = db.execute('''
-	SELECT username, code FROM user WHERE username LIKE ?
-	''', (user,))
-
-	res = cur.fetchone()
-	if res:
-		cmd.reply('%s: %s' % (res['username'], res['code']))
-	else:
-		cmd.reply('Friend code for %s could not be found.' % (user))
-
-def _user_remove(cmd):
-	cur = None
-	with db:
-		cur = db.execute('''
-		DELETE FROM user WHERE id=?
-		''', (cmd.sender['id'],))
-
-	if cur.rowcount:
-		cmd.reply('Successfully removed friend code for %s.' % (cmd.sender['username']))
-	else:
-		cmd.reply('No friend code removed. Have you registered?')
